fix: deleted SingleSummaryPage added content to DogSummaryPage
