--- src/Pages/SingleSummaryPage/SingleSummaryPage.js
@@ -1,25 +1,0 @@
-import CardWrapper from '../../Atoms/CardWrapper/CardWrapper';
-import LeaderboardList from './../../Molecules/LeaderboardList/LeaderboardList';
-import MainButton from './../../Atoms/MainButton/MainButton';
-import PenaltyOrDisqualifiedLine from '../../Atoms/Leaderboard/PenaltyOrDisqualifiedLine';
-import SummaryLine from '../../Atoms/Leaderboard/SummaryLine';
-import WYNIK_RANDOMOWEGO_PSA from './../../Data/Dummy-data/RandomDogResult';
-
-const sendDataHandler = () => {};
-
-const SingleSummaryPage = () => {
-  return (
-    <CardWrapper column>
-      <LeaderboardList result={WYNIK_RANDOMOWEGO_PSA} />
-      <PenaltyOrDisqualifiedLine result={WYNIK_RANDOMOWEGO_PSA} />
-      <SummaryLine result={WYNIK_RANDOMOWEGO_PSA} />
-      <MainButton
-        text="lista kompletna - wyślij dane"
-        secondary
-        onClick={sendDataHandler}
-      />
-    </CardWrapper>
-  );
-};
-
-export default SingleSummaryPage;
